refactor: extract Dashboard composable
